core: add transaction tags
The transaction builder should accept either transaction_tags or
reference_data, but not both. In either case, the data is saved to the
reference data column in the annotated transaction and committed (with the
new hashed tags functionality) to the raw transaction.

The resulting transaction API object should populate both the existing
reference_data field and a new tags field from the reference data column in
the annotated transaction.

When providing either transaction_tags or reference_data arguments on the
transaction builder, the data should also be copied to a new
snapshot.transaction_tags field in each action in the transaction.

https://github.com/chain/chainfork/blob/main/product/2018-03-18-transaction-tags.md

Closes #3231

Author: Howard Chen <howard7708@gmail.com>
Date: Mon Apr 9 15:27:48 2018 -0700
upstream:9f519af10e217f42621ef03747b23564727eec2b