protocol/txvm: unexport option funcs
Rather than having the option type defined as any function that can be
applied to the VM struct, make it a struct with an unexported function
that can only be called when the VM is initialized. This adds safety,
since previously options could be called at any phase of the VM's
execution.

Closes #3444

Author: Eric Rykwalder <e.rykwalder@gmail.com>
Date: Mon Apr 23 12:11:23 2018 -0700
upstream:f083769d872ade33ec922d088458544d16c9d9d0