protocol: take snapshot after transaction count, not time
Trigger snapshots based on number of transactions in the blocks committed
since our last snapshot.

Currently, we take a snapshot when a new block is created and its been over
an hour since the last time a block was created. This means that if a large
volume of transactions is processed in a short span of time and `ledgerd`
is restarted, then we have to replay a huge number of blocks from the last
snapshot before we can process new calls to `/build-transaction` and
`/submit-transaction`. By triggering snapshots on transactions rather than
time, we can avoid that latency.


Closes #3402

Author: Vicki Niu <vicki.niu@gmail.com>
Date: Wed Apr 25 16:40:04 2018 -0700
upstream:2cbc0a9c50c49ea6a1d5bb5f58adf0f41b1502bd