core/generator: simplify block signing
This PR removes the `core/blocksigner` package and coalesces `core.Signer`
and `core.BlockSigner` into a simple callback type. Most of
`generator.getAndAddBlockSignatures` is also moved to `bc.SignBlock` in
`protocol/bc`. Signed and unsigned blocks are now distinct types.

This PR also removes what looks like it was buggy code in
`getAndAddSignatures`, which expected `g.signers` to have a length equal to
the required quorum. In fact its length should be equal to the number of
pubkeys in the block predicate. We never hit the bug because our quorum and
number of pubkeys were always the same.

Finally, this PR reduces the number of times the block hash is computed for
signing to just 1.

Closes #3708

Author: Bob Glickstein <bobg@users.noreply.github.com>
Date: Thu May 17 15:23:24 2018 -0700
upstream:5251226018f704948cf8e142909ad37605e40dd7